axvirtio_blk

Struct VirtioQueue

pub struct VirtioQueue<T>
where
    T: GuestMemoryAccessor + Clone,
{
    pub index: u16,
    pub size: u16,
    pub desc_table: Option<DescriptorTable<T>>,
    pub max_size: u16,
    pub ready: bool,
    pub desc_table_addr: GuestPhysAddr,
    pub avail_ring_addr: GuestPhysAddr,
    pub used_ring_addr: GuestPhysAddr,
    pub event_idx_enabled: bool,
    /* private fields */
}

VirtIO queue implementation

Fields
index: u16

Queue index

size: u16

Queue size

desc_table: Option<DescriptorTable<T>>

Descriptor table

max_size: u16

Maximum queue size

ready: bool

Queue ready flag

desc_table_addr: GuestPhysAddr

Descriptor table address (guest physical)

avail_ring_addr: GuestPhysAddr

Available ring address (guest physical)

used_ring_addr: GuestPhysAddr

Used ring address (guest physical)

event_idx_enabled: bool

Event index enabled

Implementations

impl<T> VirtioQueue<T>
where T: GuestMemoryAccessor + Clone,

pub fn new(index: u16, size: u16, accessor: Arc<T>) -> VirtioQueue<T>

Create a new VirtIO queue

pub fn set_size(&mut self, size: u16) -> Result<(), VirtioError>

Set queue size

pub fn set_desc_table_addr( &mut self, addr: GuestPhysAddr, ) -> Result<(), VirtioError>

Set descriptor table address

pub fn set_avail_ring_addr( &mut self, addr: GuestPhysAddr, ) -> Result<(), VirtioError>

Set available ring address

pub fn set_used_ring_addr( &mut self, addr: GuestPhysAddr, ) -> Result<(), VirtioError>

Set used ring address

pub fn set_ready(&mut self, ready: bool)

Mark queue as ready

pub fn is_valid(&self) -> bool

Check if queue is valid and ready

pub fn reset(&mut self)

Reset the queue

pub fn read_avail_idx(&self) -> Result<u16, VirtioError>

Read available ring index

pub fn add_used(&mut self, desc_index: u16, len: u32) -> Result<(), VirtioError>

Add a used buffer to the used ring

pub fn pop_avail(&mut self) -> Result<Option<u16>, VirtioError>

Get next available descriptor

pub fn get_used_ring(&self) -> Option<&UsedRing<T>>

Get the used ring reference

pub fn get_used_ring_mut(&mut self) -> Option<&mut UsedRing<T>>

Get the used ring mutable reference

pub fn get_avail_ring(&self) -> Option<&AvailableRing<T>>

Get the available ring reference

pub fn get_desc_table(&self) -> Option<&DescriptorTable<T>>

Get the descriptor table reference

pub fn read_avail_entry(&self, ring_index: u16) -> Result<u16, VirtioError>

Read available ring entry

pub fn update_last_avail_idx(&mut self, idx: u16)

Update last available index

pub fn get_last_avail_idx(&self) -> u16

Get last available index

pub fn validate_virtio_block_chain( &self, head_index: u16, min_length: usize, ) -> Result<bool, VirtioError>

Validate VirtIO block chain

pub fn get_data_buffers( &self, head_index: u16, device_type: VirtioDeviceID, ) -> Result<Vec<(GuestPhysAddr, usize, bool)>, VirtioError>

Get data buffers from descriptor chain

pub fn get_status_addr( &self, head_index: u16, ) -> Result<GuestPhysAddr, VirtioError>

Get status address from descriptor chain

pub fn should_notify(&self) -> Result<bool, VirtioError>

Check if should notify

pub fn write_status_byte( &self, head_index: u16, status: u8, ) -> Result<(), VirtioError>

Write status byte to the status buffer of a descriptor chain

This method writes the status byte to the last descriptor in the chain, which should be a write-only descriptor according to VirtIO specification.